Added explicit typecast to height calculation.

+++ doc.cpp
@@ -6055,7 +6055,7 @@ void CMUSHclientDoc::SendWindowSizes (const int iNewWidth)
 
       pmyView->GetTextRect (&r);
 
-      WORD height = (r.bottom - r.top - m_iPixelOffset) / m_FontHeight;
+      WORD height = (WORD) ((r.bottom - r.top - m_iPixelOffset) / m_FontHeight);
 
       // now tell them our size
       unsigned char p [] = { IAC, SB, TELOPT_NAWS,
